Add Object Functionality
Objects can now be created and empty files initialized for them. They also
appear in the object list but insertion in the proper place is still an
issue. This also raises the question of whether objects should be saved
on an object by object basis, or if the entire project should be saved
at one go, hence files should not be created until the project is saved.