Scheme Syntax Highlighting
Since mithro asked for it...syntax highlighting has been added to
the Component panel using a wxStyledTextControl which is backed by
Scintilla. Plenty of options and opportunity there to have a lot of
functionality right from the start in terms of editing. Highlighting will
be added to the other text areas that are used for editing the TPCL code
once I play around with the STC a bit more. The code is, right now, taken
directly from the wxPython demo compilation and I'll need to give proper
credit...right now there's just a note in the file's header comment. Meh.