Ability to associate Components with Properties
It is now possible to go through the GUI to get a list of Properties
in the project that are not currently associated with a Component that
is being edited (i.e., affected by) and allows them to be added to
the Component. This should really add emphasis to the newly associated
Properties, but the current method of doing this would result in a very
inelegant solution.