Initial Code Generation Support
tpserver-cpp compiles with the new code (after I added in some of my
own debug output to see that I had left a # sign in place of a numeric
speed. Things seem a little goofy, and the way that they rulesets are
coded on the server currently relies on the order of component generation
for referencing components and properties so the minisec.cpp file itself
needs a bit of reworking, but things are looking good so far.