Object Deletion Works now
Things won't get b0rked when we delete an object. We now remove any
associations that objects have with other objects. This could be bad
if we accidentally remove something that we don't want to remove. Lots
of links to reestablish. The current method of doing things also takes
advantage of the fact that we can only delete the currently selected
object to not have to worry about updating the current edit view.