Er. Yeah. Added the TPCL evaluation code
It would help to have this around, eh? Geez. Absent minded or what,
Fro? But, on the upside, errors can now be detected in TPCL code, both
syntax errors and references to nonexistant variables or procedures. A
whole barrel of fun is opened up with that, though, since some rulesets
generate properties on the fly as designs are created, and the RDE needs
to be made aware of these. I guess this means providing a configuration
window for the project to allow the user to specify what besides the
default should be in the TPCL environment for various objects.