Silently remove invalid entries from project history
So, since I have the annoying tendency to commit partial updates and
forget important things like new folders I'm sure that everyone keeps
running into an issue of having the RDE break when they try to run
it, often because it's pointing to a project that doesn't exist. The
ConfigManager now validates each project in the project history before
adding it to the project history.