Added RDE support for multiple categories
Which raises the interesting point of how to handle category
associations. Can components only affect properties in their own
category? Should we have an implicit and hidden 'general' category in
which all objects are placed if this is the case so that basic, general
properties don't need to have their own category which would just be
annoying? Meh. Time to talk to mithro about that.