Started skeleton of TpclEditorDialog
Brief outline of functionality for each button and stuffs. Introduced
an import bug. Damn python. This has happened before and I just worked
around it. This time I'm going to get to the bottom of things. Now that
I've got the tpcl expression objects working (tested on the command line)
I'll deal with the GUI tomorrow.