Added cond expression
This is a test of the ability for expressions to be extended. It seems
to work ok. There are some issues with removed expressions that have
been added, but really if a person messes up that bad they should just
remove the entire cond expression and start again. Or an undo feature
needs to be implemented.